Moved schema to server
